Added new assets and played with pumpkin behavior
Downloaded cemetery package and created the set for the game in spooky theme.

Played with the pumpkin spawn and destroy - they now 'shuffle' away after a few seconds and the will be removed a few seconds later - to avoid spawning too many at once! (can be removed later when othere  mechanisms are employed)